Refactor hypergraph visualization to use class-based API
Replaced the procedural visualization functions in draw_hypergraph.py with a new class-based interface using IHypergraphVisualizer and HypergraphVisualizer.
Updated __init__.py to import the new visualizer classes and removes the old draw_hypergraph.py and its associated tests.
