+++ hypergraphx/viz/__init__.py
@@ -1,8 +1,6 @@
 from hypergraphx.viz.draw_communities import draw_communities
-from hypergraphx.viz.draw_hypergraph import (
-    to_nx, get_node_labels, get_pairwise_edge_labels, 
-    get_hyperedge_labels, draw, get_hyperedge_styling_data
-)
+from hypergraphx.viz.IHypergraphVisualizer import IHypergraphVisualizer
+from hypergraphx.viz.HypergraphVisualizer import HypergraphVisualizer
 from hypergraphx.viz.draw_projections import draw_bipartite, draw_clique
 from hypergraphx.viz.plot_motifs import plot_motifs
